// src/main/resources/APachas/src/app/services/entities/Product.ts
export class Product {
    productId: number;
    productName: string;
    productDescription: string;
    productCost: number;
    productQuantity: number;
    productPhoto: string;
    eventId: number;
    productPurchaseDate: string;
    productActive: boolean;
    productCreation: Date;
    productRemoval: Date;

}